UI Update
Added Instructions.webp and T_machineL.webp which is the seamless texture for the scrolling background. Added some animation to the background and made small changes in styling.Added circular clock element in countdown to use a circular clock in the javascript. Corrected a small typo in Mouse vs mouse.